[CLEAN] Volume is component + remembers old value on mute
