+++ hooks/use-player.ts
@@ -4,6 +4,7 @@ interface PlayerStore {
     ids: string[];
     source: string;
     fadeTime: number;
+    volume: number;
     activeId?: string;
     setId: (id: string) => void;
     setIds: (ids: string[], source: string) => void;
@@ -11,11 +12,13 @@ interface PlayerStore {
     bigPicture: boolean;
     toggleBigPicture: () => void;
     setFadeTime: (fadeTime: number) => void;
+    setVolume: (volume: number) => void;
 }
 
 const usePlayer = create<PlayerStore>((set, getState) => ({
     ids: [],
     source: '',
+    volume: 1,
     fadeTime: 0,
     activeId: undefined,
     setId: (id: string) => set({ activeId: id }),
@@ -23,7 +26,8 @@ const usePlayer = create<PlayerStore>((set, getState) => ({
     reset: () => set({ ids: [], activeId: undefined }),
     bigPicture: false,
     toggleBigPicture: () => set({ bigPicture: !getState().bigPicture }),
-    setFadeTime: (fadeTime: number) => set({ fadeTime: fadeTime })
+    setFadeTime: (fadeTime: number) => set({ fadeTime: fadeTime }),
+    setVolume: (volume: number) => set({ volume: volume }),
 }));
 
 export default usePlayer;
